BootStrap added, Corrected PartNumber to LONG and Body tag that was Head tag.

+++ csinventory/Migrations/20210507090341_Initial.cs
@@ -15,7 +15,7 @@ namespace csinventory.Migrations
                     Name = table.Column<string>(type: "nvarchar(max)", nullable: true),
                     Description = table.Column<string>(type: "nvarchar(max)", nullable: true),
                     Quantity = table.Column<int>(type: "int", nullable: false),
-                    PartNumber = table.Column<int>(type: "int", nullable: false),
+                    PartNumber = table.Column<long>(type: "bigint", nullable: false),
                     UnitCost = table.Column<decimal>(type: "decimal(8,2)", nullable: false),
                     TotalCost = table.Column<decimal>(type: "decimal(8,2)", nullable: false)
                 },
