+++ csinventory/Models/Part.cs
@@ -12,7 +12,7 @@ namespace csinventory.Models
         public string Name { get; set; }
         public string Description { get; set; }
         public int Quantity { get; set; }
-        public int PartNumber { get; set; }
+        public long PartNumber { get; set; }
         [Column(TypeName="decimal(8,2)")]
         public decimal UnitCost { get; set; }
         [Column(TypeName = "decimal(8,2)")]
